+++ lib/application.ex
@@ -3,6 +3,7 @@ defmodule Commanded.Application do
 
   alias Commanded.Aggregates.Aggregate
   alias Commanded.Application.Config
+  alias Commanded.Commands.Router
 
   telemetry_event(%{
     event: [:commanded, :application, :dispatch, :start],
@@ -165,6 +166,7 @@ defmodule Commanded.Application do
   """
 
   @type t :: module
+  @type options :: [name: nil | atom]
 
   @doc false
   defmacro __using__(opts) do
@@ -187,6 +189,7 @@ defmodule Commanded.Application do
         config
       end
 
+      @spec child_spec(opts :: Commanded.Application.options()) :: Supervisor.child_spec()
       def child_spec(opts) do
         %{
           id: name(opts),
@@ -205,6 +208,18 @@ defmodule Commanded.Application do
         Supervisor.stop(pid, :normal, timeout)
       end
 
+      @doc """
+      Retrieve aggregate state of an aggregate.
+
+      Retrieving aggregate state is done by calling to the opened aggregate,
+      or querying the event store for an optional state snapshot
+      and then replaying the aggregate's event stream.
+      """
+      @spec aggregate_state(
+              aggregate_module :: module(),
+              aggregate_uuid :: Aggregate.uuid(),
+              timeout :: integer
+            ) :: Aggregate.state()
       def aggregate_state(aggregate_module, aggregate_uuid, timeout \\ 5000) do
         Aggregate.aggregate_state(
           __MODULE__,
@@ -255,7 +270,7 @@ defmodule Commanded.Application do
   application is already started, or `{:error, term}` in case anything else goes
   wrong.
   """
-  @callback start_link(opts :: Keyword.t()) ::
+  @callback start_link(opts :: options) ::
               {:ok, pid}
               | {:error, {:already_started, pid}}
               | {:error, term}
@@ -272,14 +287,7 @@ defmodule Commanded.Application do
       `Commanded.Commands.Router` and included in the application.
 
   """
-  @callback dispatch(command :: struct()) ::
-              :ok
-              | {:ok, aggregate_state :: struct()}
-              | {:ok, aggregate_version :: non_neg_integer()}
-              | {:ok, execution_result :: Commanded.Commands.ExecutionResult.t()}
-              | {:error, :unregistered_command}
-              | {:error, :consistency_timeout}
-              | {:error, reason :: term()}
+  @callback dispatch(command :: struct()) :: Router.dispatch_resp()
 
   @doc """
   Dispatch a registered command.
@@ -305,10 +313,19 @@ defmodule Commanded.Application do
         - `correlation_id` - an optional UUID used to correlate related
           commands/events together.
 
-        - `consistency` - one of `:eventual` (default) or `:strong`. By
-          setting the consistency to `:strong` a successful command dispatch
-          will block until all strongly consistent event handlers and process
-          managers have handled all events created by the command.
+        - `consistency` - to choose the consistency guarantee of the command dispatch.
+
+          The available options are:
+
+          - `:eventual` (default) - a successful command dispatch will return immediately.
+
+          - `:strong` - a successful command dispatch will block until all strongly
+            consistent event handlers and process managers have handled all events created by the command.
+
+          - An explicit list of event handler and process manager modules (or their configured names),
+            containing only those handlers you'd like to wait for. No other handlers will be awaited on,
+            regardless of their own configured consistency setting.
+            e.g. `[ExampleHandler, AnotherHandler]` or `["ExampleHandler", "AnotherHandler"]`
 
         - `metadata` - an optional map containing key/value pairs comprising
           the metadata to be associated with all events created by the
@@ -350,14 +367,7 @@ defmodule Commanded.Application do
   @callback dispatch(
               command :: struct(),
               timeout_or_opts :: non_neg_integer() | :infinity | Keyword.t()
-            ) ::
-              :ok
-              | {:ok, aggregate_state :: struct()}
-              | {:ok, aggregate_version :: non_neg_integer()}
-              | {:ok, execution_result :: Commanded.Commands.ExecutionResult.t()}
-              | {:error, :unregistered_command}
-              | {:error, :consistency_timeout}
-              | {:error, reason :: term()}
+            ) :: Router.dispatch_resp()
 
   @doc false
   def dispatch(application, command, opts \\ [])