+++ lib/commanded/commands/router.ex
@@ -206,7 +206,8 @@ defmodule Commanded.Commands.Router do
 
   """
 
-  alias Commanded.Commands.Router
+  alias Commanded.Aggregates.DefaultLifespan
+  alias Commanded.Commands.{ExecutionResult, Router}
   alias Commanded.UUID
 
   defmacro __using__(opts) do
@@ -216,7 +217,7 @@ defmodule Commanded.Commands.Router do
       import unquote(__MODULE__)
 
       @before_compile unquote(__MODULE__)
-      @behaviour Commanded.Commands.Router
+      @behaviour Router
 
       Module.register_attribute(__MODULE__, :registered_commands, accumulate: true)
       Module.register_attribute(__MODULE__, :registered_middleware, accumulate: true)
@@ -227,7 +228,7 @@ defmodule Commanded.Commands.Router do
         consistency: Router.get_opt(unquote(opts), :default_consistency, :eventual),
         returning: Router.get_default_dispatch_return(unquote(opts)),
         timeout: 5_000,
-        lifespan: Commanded.Aggregates.DefaultLifespan,
+        lifespan: DefaultLifespan,
         metadata: %{},
         retry_attempts: 10
       ]
@@ -365,6 +366,15 @@ defmodule Commanded.Commands.Router do
     end
   end
 
+  @type dispatch_resp ::
+          :ok
+          | {:ok, aggregate_state :: struct()}
+          | {:ok, aggregate_version :: non_neg_integer()}
+          | {:ok, execution_result :: ExecutionResult.t()}
+          | {:error, :unregistered_command}
+          | {:error, :consistency_timeout}
+          | {:error, reason :: term()}
+
   @doc """
   Dispatch the given command to the registered handler.
 
@@ -377,14 +387,7 @@ defmodule Commanded.Commands.Router do
       :ok = BankRouter.dispatch(command)
 
   """
-  @callback dispatch(command :: struct()) ::
-              :ok
-              | {:ok, aggregate_state :: struct()}
-              | {:ok, aggregate_version :: non_neg_integer()}
-              | {:ok, execution_result :: Commanded.Commands.ExecutionResult.t()}
-              | {:error, :unregistered_command}
-              | {:error, :consistency_timeout}
-              | {:error, reason :: term()}
+  @callback dispatch(command :: struct()) :: dispatch_resp()
 
   @doc """
   Dispatch the given command to the registered handler providing a timeout.
@@ -460,14 +463,7 @@ defmodule Commanded.Commands.Router do
   @callback dispatch(
               command :: struct(),
               timeout_or_opts :: non_neg_integer() | :infinity | Keyword.t()
-            ) ::
-              :ok
-              | {:ok, aggregate_state :: struct()}
-              | {:ok, aggregate_version :: non_neg_integer()}
-              | {:ok, execution_result :: Commanded.Commands.ExecutionResult.t()}
-              | {:error, :unregistered_command}
-              | {:error, :consistency_timeout}
-              | {:error, reason :: term()}
+            ) :: dispatch_resp()
 
   defmacro __before_compile__(_env) do
     quote generated: true do