+++ lib/commanded/process_managers/process_manager.ex
@@ -60,9 +60,12 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   manager module and implement the callback functions defined in the behaviour:
 
   - `c:interested?/1`
+  - `c:interested?/2`
   - `c:handle/2`
+  - `c:handle/3`
   - `c:apply/2`
   - `c:after_command/2`
+  - `c:after_command/3`
   - `c:error/3`
 
   Please read the [Process managers](process-managers.html) guide for more
@@ -79,27 +82,41 @@ defmodule Commanded.ProcessManagers.ProcessManager do
 
         def interested?(%AnEvent{uuid: uuid}), do: {:start, uuid}
 
+        def interested?(%AnotherEvent{}, metadata),
+          do: {:continue, Map.fetch!(metadata, :correlation_id)}
+
         def handle(%ExampleProcessManager{}, %ExampleEvent{}) do
           [
             %ExampleCommand{}
           ]
         end
 
+        def handle(%ExampleProcessManager{}, %AnotherEvent{}, _metadata) do
+          [
+            %AnotherCommand{}
+          ]
+        end
+
         def after_command(%ExampleProcessManager{}, %ExampleCommand{}) do
+          :continue
+        end
+
+        def after_command(%ExampleProcessManager{}, %AnotherCommand{}, _metadata) do
           :stop
         end
 
         def error({:error, failure}, %ExampleEvent{}, _failure_context) do
           # Retry, skip, ignore, or stop process manager on error handling event
+          :skip
         end
 
         def error({:error, failure}, %ExampleCommand{}, _failure_context) do
           # Retry, skip, ignore, or stop process manager on error dispatching command
+          :skip
         end
       end
 
-  Start the process manager (or configure as a worker inside a
-  [Supervisor](supervision.html))
+  Start the process manager (or configure as a worker inside a [Supervisor](supervision.html))
 
       {:ok, process_manager} = ExampleProcessManager.start_link()
 
@@ -279,11 +296,19 @@ defmodule Commanded.ProcessManagers.ProcessManager do
 
   """
 
-  alias Commanded.ProcessManagers.FailureContext
+  alias Commanded.EventStore.RecordedEvent
 
-  @type domain_event :: struct
-  @type command :: struct
-  @type process_manager :: struct
+  alias Commanded.ProcessManagers.{
+    FailureContext,
+    ProcessManager,
+    ProcessManagerInstance,
+    ProcessRouter
+  }
+
+  @type domain_event :: struct()
+  @type enriched_metadata :: RecordedEvent.enriched_metadata()
+  @type command :: struct()
+  @type process_manager :: struct()
   @type process_uuid :: String.t() | [String.t()]
   @type consistency :: :eventual | :strong
 
@@ -299,8 +324,21 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   @doc """
   Is the process manager interested in the given command?
 
-  The `c:interested?/1` function is used to indicate which events the process
-  manager receives. The response is used to route the event to an existing
+  See `c:interested?/2` for details.
+  """
+  @callback interested?(domain_event) ::
+              {:start, process_uuid}
+              | {:start!, process_uuid}
+              | {:continue, process_uuid}
+              | {:continue!, process_uuid}
+              | {:stop, process_uuid}
+              | false
+
+  @doc """
+  Is the process manager interested in the given command?
+
+  The `c:interested?/2` function is used to indicate which events
+  the process manager receives. The response is used to route the event to an existing
   instance or start a new process instance:
 
   - `{:start, process_uuid}` - create a new instance of the process manager.
@@ -331,7 +369,7 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   problematic event.
 
   """
-  @callback interested?(domain_event) ::
+  @callback interested?(domain_event, enriched_metadata) ::
               {:start, process_uuid}
               | {:start!, process_uuid}
               | {:continue, process_uuid}
@@ -343,37 +381,62 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   Stop the process manager instance after a command is successfully
   dispatched.
 
-  The `c:after_command/2` function can be omitted if you do not need to stop
-  after a specific command or if you would instead use the `c:interested?/1`
+  See `c:after_command/3` for details.
+  """
+  @callback after_command(process_manager, command) :: :continue | :stop
+
+  @doc """
+  Stop the process manager instance after a command is successfully
+  dispatched.
+
+  The `c:after_command/3` function can be omitted if you do not need to stop
+  after a specific command or if you would instead use the `c:interested?/2`
   stop mechanism.
   """
-  @callback after_command(process_manager, domain_event) :: :continue | :stop
+  @callback after_command(process_manager, command, enriched_metadata) :: :continue | :stop
 
   @doc """
   Process manager instance handles a domain event, returning any commands to
   dispatch.
 
-  A `c:handle/2` function can be defined for each `:start` and `:continue`
-  tagged event previously specified. It receives the process manager's state and
-  the event to be handled. It must return the commands to be dispatched. This
-  may be none, a single command, or many commands.
-
-  The `c:handle/2` function can be omitted if you do not need to dispatch a
-  command and are only mutating the process manager's state.
+  See `c:handle/3` function for details.
   """
   @callback handle(process_manager, domain_event) :: command | list(command) | {:error, term}
 
   @doc """
+  Process manager instance handles a domain event, returning any commands to
+  dispatch.
+
+  A `c:handle/3` function can be defined for each `:start` and `:continue`
+  tagged event previously specified. It receives the process manager's state,
+  event to be handled, and the event's enriched metadata. It must return the
+  commands to be dispatched. This may be none, a single command, or many
+  commands.
+
+  The `c:handle/3` function can be omitted if you do not need to dispatch a
+  command and are only mutating the process manager's state.
+  """
+  @callback handle(process_manager, domain_event, enriched_metadata) ::
+              command | list(command) | {:error, term}
+
+  @doc """
   Mutate the process manager's state by applying the domain event.
 
-  The `c:apply/2` function is used to mutate the process manager's state. It
-  receives the current state and the domain event, and must return the modified
-  state.
+  See `c:apply/3` function for details.
+  """
+  @callback apply(process_manager, domain_event) :: process_manager
+
+  @doc """
+  Mutate the process manager's state by applying the domain event.
+
+  The `c:apply/3` function is used to mutate the process manager's state. It
+  receives the current state, the domain event and the event metadata, and must
+  return the modified state.
 
   This callback function is optional, the default behaviour is to retain the
   process manager's current state.
   """
-  @callback apply(process_manager, domain_event) :: process_manager
+  @callback apply(process_manager, domain_event, enriched_metadata) :: process_manager
 
   @doc """
   Called when a command dispatch or event handling returns an error.
@@ -438,20 +501,25 @@ defmodule Commanded.ProcessManagers.ProcessManager do
               | {:skip, :continue_pending}
               | {:stop, reason :: term()}
 
-  @optional_callbacks init: 1, handle: 2, apply: 2, error: 3, interested?: 1, after_command: 2
-
-  alias Commanded.ProcessManagers.ProcessManager
-  alias Commanded.ProcessManagers.ProcessRouter
+  @optional_callbacks init: 1,
+                      handle: 2,
+                      handle: 3,
+                      apply: 2,
+                      apply: 3,
+                      error: 3,
+                      interested?: 1,
+                      interested?: 2,
+                      after_command: 2,
+                      after_command: 3
 
   @doc false
-  defmacro __using__(opts) do
+  defmacro __using__(using_opts) do
     quote location: :keep do
       @before_compile unquote(__MODULE__)
       @behaviour ProcessManager
-      @opts unquote(opts)
 
       def start_link(opts \\ []) do
-        opts = Keyword.merge(@opts, opts)
+        opts = Keyword.merge(unquote(using_opts), opts)
 
         {application, name, config} = ProcessManager.parse_config!(__MODULE__, opts)
 
@@ -492,13 +560,29 @@ defmodule Commanded.ProcessManagers.ProcessManager do
     # Include default fallback functions at end, with lowest precedence
     quote generated: true do
       @doc false
-      def after_command(_process_manager, _event), do: :continue
+      def after_command(process_manager, command, _metadata),
+        do: after_command(process_manager, command)
+
+      @doc false
+      def after_command(_process_manager, _command), do: :continue
+
+      # @doc false
+      def interested?(event, _metadata),
+        do: interested?(event)
 
       @doc false
       def interested?(_event), do: false
 
       @doc false
+      def handle(process_manager, event, _metadata),
+        do: handle(process_manager, event)
+
+      @doc false
       def handle(_process_manager, _event), do: []
+
+      @doc false
+      def apply(process_manager, event, _metadata),
+        do: __MODULE__.apply(process_manager, event)
 
       @doc false
       def apply(process_manager, _event), do: process_manager
@@ -511,8 +595,8 @@ defmodule Commanded.ProcessManagers.ProcessManager do
   @doc """
   Get the identity of the current process instance.
 
-  This must only be called within a process manager's `handle/2` or `apply/2`
-  callback function.
+  This must only be called within a process manager's `handle/2`, `handle/3`, `apply/2`, or
+  `apply/3` callback functions.
 
   ## Example
 
@@ -533,7 +617,7 @@ defmodule Commanded.ProcessManagers.ProcessManager do
       end
 
   """
-  defdelegate identity(), to: Commanded.ProcessManagers.ProcessManagerInstance
+  defdelegate identity, to: ProcessManagerInstance
 
   # GenServer start options
   @start_opts [:debug, :name, :timeout, :spawn_opt, :hibernate_after]
@@ -567,6 +651,7 @@ defmodule Commanded.ProcessManagers.ProcessManager do
     end
 
     {name, config} = Keyword.pop(config, :name)
+
     name = parse_name(name)
 
     unless name do