+++ test/process_managers/support/after_command_process_manager.ex
@@ -2,7 +2,7 @@ defmodule Commanded.ProcessManagers.AfterCommandProcessManager do
   @moduledoc false
 
   alias Commanded.ProcessManagers.AfterCommandProcessManager
-  alias Commanded.ProcessManagers.ExampleAggregate.Commands.Stop
+  alias Commanded.ProcessManagers.ExampleAggregate.Commands.{Continue, Stop}
   alias Commanded.ProcessManagers.ExampleAggregate.Events.{Interested, Started}
   alias Commanded.ProcessManagers.ExampleApp
 
@@ -13,8 +13,15 @@ defmodule Commanded.ProcessManagers.AfterCommandProcessManager do
   @derive Jason.Encoder
   defstruct [:status, items: []]
 
-  def interested?(%Started{aggregate_uuid: aggregate_uuid}), do: {:start, aggregate_uuid}
-  def interested?(%Interested{aggregate_uuid: aggregate_uuid}), do: {:continue, aggregate_uuid}
+  def interested?(%Started{aggregate_uuid: aggregate_uuid}, _metadata),
+    do: {:start, aggregate_uuid}
+
+  def interested?(%Interested{aggregate_uuid: aggregate_uuid}, _metadata),
+    do: {:continue, aggregate_uuid}
+
+  def handle(%AfterCommandProcessManager{}, %Interested{index: 1, aggregate_uuid: aggregate_uuid}) do
+    %Continue{aggregate_uuid: aggregate_uuid}
+  end
 
   def handle(%AfterCommandProcessManager{}, %Interested{index: 10, aggregate_uuid: aggregate_uuid}) do
     %Stop{aggregate_uuid: aggregate_uuid}
@@ -30,7 +37,19 @@ defmodule Commanded.ProcessManagers.AfterCommandProcessManager do
     %AfterCommandProcessManager{process_manager | items: items ++ [index]}
   end
 
+  # after_command/2 callback
   def after_command(%AfterCommandProcessManager{}, %Stop{}) do
     :stop
+  end
+
+  # after_command/3 callback
+  def after_command(%AfterCommandProcessManager{}, %Continue{}, metadata) do
+    %{"notify_to" => notify_to} = metadata
+
+    notify_to
+    |> :erlang.list_to_pid()
+    |> Process.send(:metadata_available, [])
+
+    :continue
   end
 end