+++ test/process_managers/support/routing/routing_process_manager.ex
@@ -8,6 +8,10 @@ defmodule Commanded.ProcessManagers.RoutingProcessManager do
     defstruct [:process_uuid, :reply_to, strict?: false]
   end
 
+  defmodule StartedFromMetadata do
+    defstruct [:process_uuid, :reply_to, strict?: false]
+  end
+
   defmodule Continued do
     @enforce_keys [:process_uuid]
     defstruct [:process_uuid, :reply_to, strict?: false]
@@ -28,6 +32,9 @@ defmodule Commanded.ProcessManagers.RoutingProcessManager do
 
   defstruct [:processes]
 
+  def interested?(%StartedFromMetadata{}, %{"process_uuid" => process_uuid} = _metadata),
+    do: {:start, process_uuid}
+
   def interested?(%Started{process_uuid: process_uuid, strict?: true}),
     do: {:start!, process_uuid}
 
@@ -42,6 +49,14 @@ defmodule Commanded.ProcessManagers.RoutingProcessManager do
   def interested?(%Stopped{process_uuid: process_uuid}), do: {:stop, process_uuid}
 
   def interested?(%Errored{}), do: raise("error")
+
+  def handle(%RoutingProcessManager{}, %StartedFromMetadata{} = event) do
+    %StartedFromMetadata{reply_to: reply_to} = event
+
+    send(reply_to, {:started, self()})
+
+    []
+  end
 
   def handle(%RoutingProcessManager{}, %Started{} = event) do
     %Started{reply_to: reply_to} = event